\name{orsk-internal}
\title{Internal orsk functions}
\alias{multiStartNew}
\description{Internal orsk functions}
\author{Zhu Wang}
\details{Internal function to be used in the main function \code{orsk}, and not intended for use by users. \code{multiStartNew} is a (slightly) modified version of multiStart in the R package BB, to suppress printout.}

\keyword{internal}                  
