\name{orsk-package}
\alias{orsk-package}
\docType{package}
\title{
Converting Odds Ratio to Relative Risk in Cohort Studies with Partial Data Information}
\description{
Concertinaing Odds Ratio to Relative Risk in Cohort Studies with Partial Data Information
}
\details{
\tabular{ll}{
Package: \tab orsk\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2013-10-08\cr
License: \tab GPL 2.0 or Above\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhu Wang

Maintainer: Zhu Wang <zwang@ccmckids.org>
}
\keyword{ package }
