\name{orsk-package}
\alias{orsk-package}
\docType{package}
\title{
Convertinging Odds Ratio to Relative Risk with Partial Data Information (Estimating Relative Risk)
}
\description{
Convertinging Odds Ratio to Relative Risk with Partial Data Information
}
\details{
\tabular{ll}{
Package: \tab orsk\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-12-22\cr
License: \tab GPL 2.0 or Above\cr
LazyLoad: \tab yes\cr
}
}
\author{
Zhu Wang

Maintainer: Zhu Wang <zwang@ccmckids.org>
}
\keyword{ package }
